requirements.txt
asgiref==3.3.4
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.25.0
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.4
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
idna==3.6
oauthlib==3.1.1
Pillow==8.2.0
psycopg2==2.9.1
pycparser==2.21
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.31.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==2.1.0
